Move fileNameFromUrl() to IOUtil
